Updated code to Oqtane version 3.1 and revised documentation navigation with new sections
Renamed Articles top category to Overview and added a concept about the hierarchical approach of Oqtane. Also added a section for video links.
For the Administration documentation, added a section about navigation and revised the left main menu and updated many of the descriptions. Also added a banner for the top level of the navigation items in the documentation.